requirements.txt
absl-py==1.4.0
altair==5.0.1
ansi2html==1.8.0
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
category-encoders==2.6.1
certifi==2023.5.7
charset-normalizer==3.1.0
clang==5.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
dacite==1.8.1
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.14
deprecation==2.1.0
dython==0.7.4
fastjsonschema==2.17.1
FLAML==1.2.4
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.1.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.10.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipywidgets==8.0.7
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter-dash==0.4.2
jupyterlab-widgets==3.0.8
kaleido==0.2.1
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
lightgbm==3.3.5
llvmlite==0.40.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
multimethod==1.9.1
nbformat==5.9.0
networkx==3.1
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.1
pandas==1.5.3
pandas-profiling==3.6.6
patsy==0.5.3
phik==0.12.3
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.15.0
plotly-resampler==0.8.3.2
pmdarima==2.0.3
protobuf==4.23.3
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycaret==3.0.4
pydantic==1.10.11
pydeck==0.8.1b0
Pympler==1.0.1
pyod==1.1.0
pyparsing==3.1.0
pyrsistent==0.19.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.4.2
rsa==4.9
schemdraw==0.15
scikit-base==0.5.0
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
six==1.15.0
sktime==0.20.0
smmap==5.0.0
statsmodels==0.14.0
streamlit==1.23.1
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tqdm==4.65.0
trace-updater==0.0.9.1
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
visions==0.7.5
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wordcloud==1.9.2
wrapt==1.12.1
wurlitzer==3.0.3
xgboost==1.7.6
xxhash==3.2.0
ydata-profiling==4.3.1
yellowbrick==1.5
